refactor agent files
