+++ backend/agent_swarm/llama_workflow/data_loading_agent.py
@@ -1,0 +1,114 @@
+"""
+Data Loading Task Agent Module
+
+This module defines the data loading task agent used in LlamaIndex agent workflows.
+"""
+
+from typing import Any, Dict, Optional
+import logging
+
+from llama_index.core.llms import LLM
+
+from .base import BaseTaskAgent
+
+
+class DataLoadingTaskAgent(BaseTaskAgent):
+    """
+    Task agent for loading data from various sources.
+
+    Responsibilities:
+    - Loading data from files (CSV, Excel, JSON, etc.)
+    - Loading data from databases
+    - Loading data from APIs
+    - Basic data validation
+    """
+
+    def __init__(self, llm: Optional[LLM] = None):
+        """Initialize the DataLoadingTaskAgent."""
+        super().__init__(name="DataLoadingAgent", llm=llm)
+
+    def run(self, input_data: Dict[str, Any]) -> Dict[str, Any]:
+        """
+        Execute the data loading task.
+
+        Args:
+            input_data: Input data for the task
+
+        Returns:
+            Dict containing loaded data and metadata
+        """
+        environment = input_data.get("environment", {})
+        goals = input_data.get("goals", [])
+
+        # Get data sources from the environment
+        data_sources = environment.get("Data", {})
+        specific_sources = input_data.get("data_sources", {})
+
+        # If specific sources are provided, use those instead
+        if specific_sources:
+            data_sources = specific_sources
+
+        self.logger.info(f"Loading data from {len(data_sources)} sources")
+
+        loaded_data = {}
+        schema_info = {}
+
+        # Process each data source
+        for source_name, source_info in data_sources.items():
+            try:
+                source_type = source_info.get("type", "unknown")
+                source_path = source_info.get("path", "")
+
+                if source_type == "csv":
+                    # Simulate loading CSV data
+                    self.logger.info(f"Loading CSV data from {source_path}")
+                    loaded_data[source_name] = {
+                        "type": "dataframe",
+                        "rows": 100,  # Simulated row count
+                        "source": source_path,
+                    }
+                    schema_info[source_name] = {
+                        "columns": ["column1", "column2", "column3"],
+                        "types": ["int", "string", "float"],
+                    }
+                elif source_type == "excel":
+                    # Simulate loading Excel data
+                    self.logger.info(f"Loading Excel data from {source_path}")
+                    loaded_data[source_name] = {
+                        "type": "dataframe",
+                        "rows": 200,  # Simulated row count
+                        "source": source_path,
+                    }
+                    schema_info[source_name] = {
+                        "columns": ["column1", "column2", "column3"],
+                        "types": ["int", "string", "float"],
+                    }
+                elif source_type == "database":
+                    # Simulate loading database data
+                    connection_string = source_info.get("connection", "")
+                    query = source_info.get("query", "")
+                    self.logger.info(f"Loading database data from {connection_string}")
+                    loaded_data[source_name] = {
+                        "type": "dataframe",
+                        "rows": 500,  # Simulated row count
+                        "source": f"DB: {connection_string}",
+                    }
+                    schema_info[source_name] = {
+                        "columns": ["id", "name", "value", "date"],
+                        "types": ["int", "string", "float", "datetime"],
+                    }
+                else:
+                    self.logger.warning(f"Unknown data source type: {source_type}")
+                    loaded_data[source_name] = {
+                        "type": "unknown",
+                        "error": f"Unsupported data source type: {source_type}",
+                    }
+            except Exception as e:
+                self.logger.error(f"Error loading data from {source_name}: {str(e)}")
+                loaded_data[source_name] = {"type": "error", "error": str(e)}
+
+        # Return the loaded data
+        return {
+            "Data Overview.raw_data": loaded_data,
+            "Data Overview.schema": schema_info,
+        }
