+++ backend/agent_swarm/llama_workflow/exploration_agent.py
@@ -1,0 +1,129 @@
+"""
+Exploration Task Agent Module
+
+This module defines the exploration task agent used in LlamaIndex agent workflows.
+"""
+
+from typing import Any, Dict, Optional
+import logging
+
+from llama_index.core.llms import LLM
+
+from .base import BaseTaskAgent
+
+
+class ExplorationTaskAgent(BaseTaskAgent):
+    """
+    Task agent for exploring and understanding data.
+
+    Responsibilities:
+    - Data profiling
+    - Statistical analysis
+    - Feature discovery
+    - Correlation analysis
+    """
+
+    def __init__(self, llm: Optional[LLM] = None):
+        """Initialize the ExplorationTaskAgent."""
+        super().__init__(name="ExplorationAgent", llm=llm)
+
+    def run(self, input_data: Dict[str, Any]) -> Dict[str, Any]:
+        """
+        Execute the data exploration task.
+
+        Args:
+            input_data: Input data for the task
+
+        Returns:
+            Dict containing exploration results
+        """
+        environment = input_data.get("environment", {})
+        goals = input_data.get("goals", [])
+
+        # Get raw data from the environment
+        data_overview = environment.get("Data Overview", {})
+        raw_data = data_overview.get("raw_data", {})
+        schema = data_overview.get("schema", {})
+
+        self.logger.info("Exploring data")
+
+        # Initialize results
+        summary = {}
+        statistics = {}
+
+        # Process each dataset
+        for dataset_name, dataset_info in raw_data.items():
+            try:
+                # Skip datasets with errors
+                if dataset_info.get("type") == "error":
+                    self.logger.warning(
+                        f"Skipping dataset {dataset_name} due to loading error"
+                    )
+                    continue
+
+                # Get dataset schema
+                dataset_schema = schema.get(dataset_name, {})
+                columns = dataset_schema.get("columns", [])
+
+                # Generate summary for this dataset
+                summary[dataset_name] = {
+                    "row_count": dataset_info.get("rows", 0),
+                    "column_count": len(columns),
+                    "columns": columns,
+                    "completeness": "95%",  # Simulated completeness
+                    "has_missing_values": True,  # Simulated missing values check
+                    "has_duplicates": False,  # Simulated duplicates check
+                }
+
+                # Generate statistics for this dataset
+                statistics[dataset_name] = {
+                    "numeric_columns": {},
+                    "categorical_columns": {},
+                    "datetime_columns": {},
+                }
+
+                # Simulate statistics for each column based on its type
+                for i, col in enumerate(columns):
+                    col_type = (
+                        dataset_schema.get("types", [])[i]
+                        if i < len(dataset_schema.get("types", []))
+                        else "unknown"
+                    )
+
+                    if col_type in ["int", "float", "double", "numeric"]:
+                        # Numeric column statistics
+                        statistics[dataset_name]["numeric_columns"][col] = {
+                            "min": 0,  # Simulated min
+                            "max": 100,  # Simulated max
+                            "mean": 50,  # Simulated mean
+                            "median": 48,  # Simulated median
+                            "std": 15,  # Simulated standard deviation
+                            "missing": 2,  # Simulated missing count
+                        }
+                    elif col_type in ["string", "text", "varchar"]:
+                        # Categorical column statistics
+                        statistics[dataset_name]["categorical_columns"][col] = {
+                            "unique_values": 10,  # Simulated unique count
+                            "most_common": [
+                                "value1",
+                                "value2",
+                                "value3",
+                            ],  # Simulated most common values
+                            "missing": 1,  # Simulated missing count
+                        }
+                    elif col_type in ["date", "datetime", "timestamp"]:
+                        # Datetime column statistics
+                        statistics[dataset_name]["datetime_columns"][col] = {
+                            "min_date": "2020-01-01",  # Simulated min date
+                            "max_date": "2023-12-31",  # Simulated max date
+                            "missing": 0,  # Simulated missing count
+                        }
+            except Exception as e:
+                self.logger.error(f"Error exploring dataset {dataset_name}: {str(e)}")
+                summary[dataset_name] = {"error": str(e)}
+
+        # Return the exploration results
+        return {
+            "Data Overview.summary": summary,
+            "Data Overview.statistics": statistics,
+        }
